+++ indico/modules/events/registration/api.py
@@ -14,6 +14,7 @@ from werkzeug.exceptions import BadRequest, Forbidden
 from indico.core import signals
 from indico.modules.events.controllers.base import RHProtectedEventBase
 from indico.modules.events.models.events import Event
+from indico.modules.events.registration.models.registrations import RegistrationState
 from indico.modules.events.registration.util import build_registration_api_data, build_registrations_api_data
 from indico.modules.oauth import oauth
 from indico.web.rh import RH
@@ -49,6 +50,8 @@ class RHAPIRegistrant(RH):
             raise BadRequest("Invalid fields: {}".format(', '.join(invalid_fields)))
 
         if 'checked_in' in request.json:
+            if self._registration.state not in (RegistrationState.complete, RegistrationState.unpaid):
+                raise BadRequest('This registration cannot be marked as checked-in')
             self._registration.checked_in = bool(request.json['checked_in'])
             signals.event.registration_checkin_updated.send(self._registration)